refactor: Switch references of ServerlessError from 'lib/classes/Error' to 'lib/serverless-error'
