+++ lib/plugins/aws/configCredentials.js
@@ -43,9 +43,7 @@ class AwsConfigCredentials {
     };
 
     if (!os.homedir()) {
-      throw new ServerlessError(
-        "Can't find home directory on your local file system."
-      );
+      throw new ServerlessError("Can't find home directory on your local file system.");
     }
 
     this.hooks = {
