+++ lib/plugins/aws/customResources/index.js
@@ -51,9 +51,7 @@ async function addCustomResourceToService(awsProvider, resourceName, iamRoleStat
     Handler = 'apiGatewayCloudWatchRole/handler.handler';
     customResourceFunctionLogicalId = awsProvider.naming.getCustomResourceApiGatewayAccountCloudWatchRoleHandlerFunctionLogicalId();
   } else {
-    throw new ServerlessError(
-      `No implementation found for Custom Resource "${resourceName}"`
-    );
+    throw new ServerlessError(`No implementation found for Custom Resource "${resourceName}"`);
   }
   absoluteFunctionName = `${funcPrefix}-${functionName}`;
   if (absoluteFunctionName.length > 64) {
