+++ lib/plugins/aws/deploy/lib/existsDeploymentBucket.js
@@ -23,9 +23,7 @@ module.exports = {
         return BbPromise.resolve();
       })
       .catch((err) => {
-        throw new ServerlessError(
-          `Could not locate deployment bucket. Error: ${err.message}`
-        );
+        throw new ServerlessError(`Could not locate deployment bucket. Error: ${err.message}`);
       });
   },
 };
