+++ lib/plugins/aws/invoke.js
@@ -29,10 +29,7 @@ class AwsInvoke {
       return await this.serverless.utils.readFile(absolutePath);
     } catch (err) {
       if (err.code === 'ENOENT') {
-        throw new ServerlessError(
-          'The file you provided does not exist.',
-          'FILE_NOT_FOUND'
-        );
+        throw new ServerlessError('The file you provided does not exist.', 'FILE_NOT_FOUND');
       }
       throw err;
     }
