+++ lib/plugins/aws/invokeLocal/index.js
@@ -67,9 +67,7 @@ class AwsInvokeLocal {
     const exists = await fileExists(absolutePath);
 
     if (!exists) {
-      throw new ServerlessError(
-        `The file you provided does not exist: ${absolutePath}`
-      );
+      throw new ServerlessError(`The file you provided does not exist: ${absolutePath}`);
     }
     if (absolutePath.endsWith('.js')) {
       // to support js - export as an input data
