+++ lib/plugins/aws/lib/validate.js
@@ -5,9 +5,7 @@ const ServerlessError = require('../../../serverless-error');
 module.exports = {
   validate() {
     if (!this.serverless.config.servicePath) {
-      throw new ServerlessError(
-        'This command can only be run inside a service directory'
-      );
+      throw new ServerlessError('This command can only be run inside a service directory');
     }
 
     this.options.stage = this.provider.getStage();
