+++ lib/plugins/aws/package/compile/events/alb/lib/validate.js
@@ -106,9 +106,7 @@ module.exports = {
     }
     const matches = listenerArn.match(this.ALB_LISTENER_REGEXP);
     if (!matches) {
-      throw new ServerlessError(
-        `Invalid ALB listenerArn in function "${functionName}".`
-      );
+      throw new ServerlessError(`Invalid ALB listenerArn in function "${functionName}".`);
     }
     return { albId: matches[1], listenerId: matches[2] };
   },
