+++ lib/plugins/aws/package/compile/events/apiGateway/lib/restApi.js
@@ -29,9 +29,7 @@ module.exports = {
         vpcEndpointIds = this.serverless.service.provider.vpcEndpointIds;
 
         if (endpointType !== 'PRIVATE') {
-          throw new ServerlessError(
-            'VPC endpoint IDs are only available for private APIs'
-          );
+          throw new ServerlessError('VPC endpoint IDs are only available for private APIs');
         }
       }
     }
