+++ lib/plugins/aws/package/compile/functions.js
@@ -382,10 +382,7 @@ class AwsCompileFunctions {
           'ensure that function has vpc configured ',
           'on function or provider level',
         ].join('');
-        throw new ServerlessError(
-          errorMessage,
-          'LAMBDA_FILE_SYSTEM_CONFIG_MISSING_VPC'
-        );
+        throw new ServerlessError(errorMessage, 'LAMBDA_FILE_SYSTEM_CONFIG_MISSING_VPC');
       }
 
       const iamRoleLambdaExecution = cfTemplate.Resources.IamRoleLambdaExecution;
