+++ lib/plugins/plugin/lib/utils.js
@@ -11,9 +11,7 @@ const ServerlessError = require('../../../serverless-error');
 module.exports = {
   validate() {
     if (!this.serverless.config.servicePath) {
-      throw new ServerlessError(
-        'This command can only be run inside a service directory'
-      );
+      throw new ServerlessError('This command can only be run inside a service directory');
     }
 
     return BbPromise.resolve();
@@ -21,9 +19,7 @@ module.exports = {
 
   getServerlessFilePath() {
     if (this.serverless.configurationPath) return this.serverless.configurationPath;
-    throw new ServerlessError(
-      'Could not find any serverless service definition file.'
-    );
+    throw new ServerlessError('Could not find any serverless service definition file.');
   },
 
   getPlugins() {
