+++ test/unit/lib/Serverless.test.js
@@ -13,7 +13,7 @@ const Utils = require('../../../lib/classes/Utils');
 const Service = require('../../../lib/classes/Service');
 const ConfigSchemaHandler = require('../../../lib/classes/ConfigSchemaHandler');
 const CLI = require('../../../lib/classes/CLI');
-const { ServerlessError } = require('../../../lib/classes/Error');
+const ServerlessError = require('../../../lib/serverless-error');
 const runServerless = require('../../utils/run-serverless');
 const fixtures = require('../../fixtures');
 const fs = require('fs');
