+++ test/unit/lib/plugins/aws/deploy/lib/createStack.test.js
@@ -6,7 +6,7 @@ const path = require('path');
 const AwsProvider = require('../../../../../../../lib/plugins/aws/provider');
 const AwsDeploy = require('../../../../../../../lib/plugins/aws/deploy/index');
 const Serverless = require('../../../../../../../lib/Serverless');
-const { ServerlessError } = require('../../../../../../../lib/classes/Error');
+const ServerlessError = require('../../../../../../../lib/serverless-error');
 const { getTmpDirPath } = require('../../../../../../utils/fs');
 
 chai.use(require('chai-as-promised'));
