+++ test/unit/lib/plugins/config.test.js
@@ -6,7 +6,7 @@ const os = require('os');
 const BbPromise = require('bluebird');
 const { expect } = require('chai');
 const config = require('@serverless/utils/config');
-const { ServerlessError } = require('../../../../lib/classes/Error');
+const ServerlessError = require('../../../../lib/serverless-error');
 const runServerless = require('../../../utils/run-serverless');
 const isTabCompletionSupported = require('../../../../lib/utils/tabCompletion/isSupported');
